--- app/components/UI/Predict/hooks/usePredictActiveOrder.ts
@@ -1,6 +1,0 @@
-import { useSelector } from 'react-redux';
-import { selectPredictActiveOrder } from '../selectors/predictController';
-
-export function usePredictActiveOrder() {
-  return useSelector(selectPredictActiveOrder);
-}
