+++ app/components/UI/Predict/hooks/usePredictDepositAndOrder.ts
@@ -1,27 +1,40 @@
-import { NavigationProp, useNavigation } from '@react-navigation/native';
-import { useCallback, useContext } from 'react';
+import {
+  NavigationProp,
+  StackActions,
+  useNavigation,
+} from '@react-navigation/native';
+import {
+  useCallback,
+  useContext,
+  useRef,
+  useState,
+  type MutableRefObject,
+} from 'react';
 import { useSelector } from 'react-redux';
 import { strings } from '../../../../../locales/i18n';
 import { IconName } from '../../../../component-library/components/Icons/Icon';
 import { ToastContext } from '../../../../component-library/components/Toast';
 import { ToastVariants } from '../../../../component-library/components/Toast/Toast.types';
+import Routes from '../../../../constants/navigation/Routes';
 import Engine from '../../../../core/Engine';
 import Logger from '../../../../util/Logger';
 import { useAppThemeFromContext } from '../../../../util/theme';
+import useApprovalRequest from '../../../Views/confirmations/hooks/useApprovalRequest';
 import { PREDICT_CONSTANTS } from '../constants/errors';
+import { OrderPreview, PlaceOrderParams } from '../providers/types';
 import { selectPredictPendingDepositByAddress } from '../selectors/predictController';
 import {
   PredictBuyPreviewParams,
   PredictNavigationParamList,
 } from '../types/navigation';
-import { PlaceOrderParams } from '../providers/types';
 import { getEvmAccountFromSelectedAccountGroup } from '../utils/accounts';
 import { ensureError } from '../utils/predictErrorHandler';
 import { usePredictConfirmNavigation } from './usePredictConfirmNavigation';
+import { usePredictPaymentToken } from './usePredictPaymentToken';
 import { usePredictTokenSelection } from './usePredictTokenSelection';
 import { usePredictTrading } from './usePredictTrading';
 
-interface PredictDepositAndOrderParams {
+export interface PredictDepositAndOrderParams {
   amountUsd?: number;
   isInputFocused?: boolean;
   transactionError?: string;
@@ -33,22 +46,56 @@ interface PredictDepositAndOrderParams {
 
 interface UsePredictDepositAndOrderParams {
   tokenSelectionParams?: PredictDepositAndOrderParams;
+  market?: PredictBuyPreviewParams['market'];
+  outcome?: PredictBuyPreviewParams['outcome'];
+  outcomeToken?: PredictBuyPreviewParams['outcomeToken'];
+  orderAmountUsd?: number;
+  depositTransactionId?: string;
+  preview?: OrderPreview | null;
+}
+
+interface UsePredictDepositAndOrderResult {
+  depositAndOrder: (params: PredictDepositAndOrderParams) => Promise<void>;
+  isDepositPending: boolean;
+  shouldPreserveActiveOrderOnUnmountRef: MutableRefObject<boolean>;
+  isDepositAndOrderLoading: boolean;
+  isConfirming: boolean;
+  confirmError?: string;
+  handleConfirm: () => Promise<void>;
 }
 
 export const usePredictDepositAndOrder = ({
   tokenSelectionParams,
-}: UsePredictDepositAndOrderParams = {}) => {
+  market,
+  outcome,
+  outcomeToken,
+  orderAmountUsd,
+  depositTransactionId,
+  preview,
+}: UsePredictDepositAndOrderParams = {}): UsePredictDepositAndOrderResult => {
   const { navigateToConfirmation } = usePredictConfirmNavigation();
   const theme = useAppThemeFromContext();
   const { toastRef } = useContext(ToastContext);
   const navigation =
     useNavigation<NavigationProp<PredictNavigationParamList>>();
+  const { onConfirm: onApprovalConfirm } = useApprovalRequest();
+  const { isPredictBalanceSelected } = usePredictPaymentToken();
 
   const evmAccount = getEvmAccountFromSelectedAccountGroup();
   const selectedInternalAccountAddress = evmAccount?.address ?? '0x0';
 
-  const { depositAndOrder: depositAndOrderWithConfirmation } =
+  const { depositAndOrder: depositAndOrderWithConfirmation, placeOrder } =
     usePredictTrading();
+  const normalizedOrderAmountUsd = orderAmountUsd ?? 0;
+  const hasExecutionParams = Boolean(
+    market || outcome || outcomeToken || preview,
+  );
+  const marketId = market?.id;
+  const outcomeId = outcome?.id;
+  const previewRef = useRef(preview);
+  previewRef.current = preview;
+  const [isConfirming, setIsConfirming] = useState(false);
+  const [confirmError, setConfirmError] = useState<string>();
 
   const depositBatchId = useSelector(
     selectPredictPendingDepositByAddress({
@@ -97,6 +144,109 @@ export const usePredictDepositAndOrder = ({
       toastRef,
     ],
   );
+
+  const showOrderPlacedToast = useCallback(() => {
+    toastRef?.current?.showToast({
+      variant: ToastVariants.Icon,
+      iconName: IconName.Check,
+      labelOptions: [
+        { label: strings('predict.order.prediction_placed'), isBold: true },
+      ],
+      hasNoTimeout: false,
+    });
+  }, [toastRef]);
+
+  const redirectToBuyPreviewForAutoOrder = useCallback(() => {
+    if (!market || !outcome || !outcomeToken || normalizedOrderAmountUsd <= 0) {
+      setConfirmError(strings('predict.deposit.error_description'));
+      setIsConfirming(false);
+      return;
+    }
+
+    setIsConfirming(false);
+    navigation.dispatch(
+      StackActions.replace(Routes.PREDICT.MODALS.BUY_PREVIEW, {
+        market,
+        outcome,
+        outcomeToken,
+        amount: normalizedOrderAmountUsd,
+        transactionId: depositTransactionId,
+        animationEnabled: false,
+      }),
+    );
+  }, [
+    depositTransactionId,
+    market,
+    normalizedOrderAmountUsd,
+    outcome,
+    outcomeToken,
+    navigation,
+  ]);
+
+  const handleOrderSuccess = useCallback(async () => {
+    const latestPreview = previewRef.current;
+    if (!latestPreview) {
+      setIsConfirming(false);
+      return;
+    }
+
+    try {
+      await placeOrder({
+        preview: latestPreview,
+        analyticsProperties: {
+          marketId,
+          outcome: outcomeId,
+        },
+      });
+      showOrderPlacedToast();
+    } catch (err) {
+      setConfirmError(
+        err instanceof Error
+          ? err.message
+          : strings('predict.deposit.error_description'),
+      );
+    } finally {
+      setIsConfirming(false);
+      navigation.goBack();
+    }
+  }, [marketId, navigation, outcomeId, placeOrder, showOrderPlacedToast]);
+
+  const handleConfirm = useCallback(async () => {
+    if (!hasExecutionParams || isConfirming) {
+      return;
+    }
+
+    setIsConfirming(true);
+    setConfirmError(undefined);
+
+    if (isPredictBalanceSelected) {
+      await handleOrderSuccess();
+      return;
+    }
+
+    try {
+      redirectToBuyPreviewForAutoOrder();
+      await onApprovalConfirm({
+        deleteAfterResult: true,
+        waitForResult: true,
+        handleErrors: false,
+      });
+    } catch (err) {
+      setConfirmError(
+        err instanceof Error
+          ? err.message
+          : strings('predict.deposit.error_description'),
+      );
+      setIsConfirming(false);
+    }
+  }, [
+    hasExecutionParams,
+    handleOrderSuccess,
+    isConfirming,
+    isPredictBalanceSelected,
+    onApprovalConfirm,
+    redirectToBuyPreviewForAutoOrder,
+  ]);
 
   const depositAndOrder = useCallback(
     async (params: PredictDepositAndOrderParams) => {
@@ -173,5 +323,8 @@ export const usePredictDepositAndOrder = ({
     isDepositPending: !!depositBatchId,
     shouldPreserveActiveOrderOnUnmountRef,
     isDepositAndOrderLoading,
+    isConfirming,
+    confirmError,
+    handleConfirm,
   };
 };
